Fixed some bugs with input file size checking.

+++ src/core/Options.h
@@ -32,7 +32,7 @@ public:
 	std::string tcpflowParams;
 	int emulationSteps;
 	bool skipBigFiles;
-	int bigFileSize;
+	unsigned long bigFileSize;
 	bool skipBrokenSamples;
 	int brokenSampleSize;
